Broke out the createElement function into a new file to be imported into main.js. #2
