+++ client/utils/createElement.js
@@ -1,0 +1,13 @@
+function createElement(tagName, attributes, content, $children) {
+  const $element = document.createElement(tagName)
+  $element.textContent = content
+  for (const key in attributes) {
+    $element.setAttribute(key, attributes[key])
+  }
+  $children.forEach($child => {
+    $element.appendChild($child)
+  })
+  return $element
+}
+
+module.exports = createElement
